+++ tests/TestCase.php
@@ -3,6 +3,7 @@
 namespace DrPshtiwan\LivewireMediaSelector\Tests;
 
 use DrPshtiwan\LivewireMediaSelector\MediaSelectorServiceProvider;
+use Livewire\LivewireServiceProvider;
 use Orchestra\Testbench\TestCase as BaseTestCase;
 
 abstract class TestCase extends BaseTestCase
@@ -10,7 +11,7 @@ abstract class TestCase extends BaseTestCase
     protected function getPackageProviders($app)
     {
         return [
-            \Livewire\LivewireServiceProvider::class,
+            LivewireServiceProvider::class,
             MediaSelectorServiceProvider::class,
         ];
     }